RFC: JModels.jl

This document specifies an interface for statistical Julia packages. The goal is to create a generic interface for many (wildly) different packages.

One main goal of this interface is as follows. Suppose that there is a Julia cross-validation package called CV.jl satisfying the JModels interface. This interface should make it possible to use CV.jl on models defined by organisations such as

MLJ

SciML

Flux

Turing

JuliaStats

JuliaML

Invenia

and more.

Related work

LearningStrategies.jl provides an abstract interface for iteratively training a model. Specifically, the package allows for a model setup!, iteratively update! and a cleanup!. It has been the foundation for IterationControl.jl.

Model definition

Let $R^k$ denote $k$ random variables. We define a statistical model $m$ with parameters $p$ as a mapping from $v$ random variables to $w$ random variables:

\[m_p(X :: R^v) :: R^w,\]

where $1 \leq v$ and $1 \leq w$ and $X$ satisfies the assumptions of $m_p$. Here, $m_p$ is often called a trained model. Let such a model be obtained by passing $u$ random variables to $m$:

\[m_p = m(T :: R^u).\]

This is often called training a model.

Data definition

This interface makes no assumptions about the datatype. It is up to the package who implements the interface to decide what datatypes are allowed.

Note that, in the case of statistical models, the Tables.jl interface is not generic enough. For example, for image classifiers, the data cannot easily be contained in a table.

API Reference

All the methods listed below are considered public, that is, can be used or extended.

JModels.fit — Function

JModels.fit(t, x; kwargs...)

Return a fitted model of type t on data. Implementing this function is encouraged but optional. Without implementing this function, things such as model evaluation where your model is fitted and applyd multiple times are not possible.

It is advised to assign default values to all keyword arguments. This makes it easier for people to compare different models.

JModels.fit! — Function

JModels.fit!(model, data; kwargs...)

Fit an existing model on data by mutating model. In contrast to fit, this method is more flexible in configuring the model since a predefined model can be passed to be fitted. Also, this method can offer more performance if the model is trained in multiple steps.

JModels.predict — Function

JModels.predict(model, data; kwargs...)

Predict with model on data. For example, a k-means clustering model can predict target labels.

JModels.transform — Function

JModels.transform(model, data; kwargs...)

Transform data via model. For example, a k-means clustering model reduce dimensionality.

JModels.inverse_transform — Function

JModels.inverse_transform(model, data; kwargs...)

Inversely transform data via model.

JModels.ismodel — Function

JModels.ismodel(x) -> Bool

Check if an object x has defined that it is a statistical model and has implemented the JModels interface.

JModels.verify_model — Function

JModels.verify_model(x)

Throw an error if !ismodel(x).

Model evaluation

As an example, this section discusses how cross-validation (CV) could be applied to different Julia models via this interface. CV for labeled data can be defined as evaluating the model $m_p$ for $k$ datasets, that is, $k$ sets of random variables $X_i :: R^v$. Specifically, it requires the collection

\[Y = [ m_p(X_i :: R^v) \text{ for } i \text{ in } 1:k ],\]

which can then be evaluated by applying an scoring function $\text{score}$ to each outcome $Y_i \in Y$:

\[\text{aggregate}([ \text{score}(Y_i) \text{ for } Y_i \text{ in } Y ]).\]

For a collection of model types MT and data, this could be implemented via something like (yes, an actual implementation is needed to verify correctness):

function cross_validate(mt, data)
    return map(train_test_splits(data)) do (train, test)
        model = fit(mt, train)
        predictions = apply(model, test.x)
        root_mean_squared_error(test.x, test.y)
    end
end

scores = [aggregate(cross_validate(mt, data)) for mt in MT]